Documentation and refactors
